Feature/project (#24)

+++ Task-api/src/main/java/com/nhnacademy/taskapi/service/milestone/impl/MileStoneServiceImpl.java
@@ -7,14 +7,15 @@ import com.nhnacademy.taskapi.entity.milestone.request.MileStoneRequest;
 import com.nhnacademy.taskapi.error.milestone.MileStoneAlreadyExistsException;
 import com.nhnacademy.taskapi.error.milestone.MileStoneNotFoundException;
 import com.nhnacademy.taskapi.error.project.ProjectNotFoundException;
+import com.nhnacademy.taskapi.error.projectmember.ProjectMemberUserNotFoundException;
 import com.nhnacademy.taskapi.repository.milestone.MileStoneRepository;
 import com.nhnacademy.taskapi.repository.project.ProjectRepository;
+import com.nhnacademy.taskapi.repository.projectmember.ProjectMemberRepository;
 import com.nhnacademy.taskapi.service.milestone.MileStoneService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -23,6 +24,7 @@ import java.util.Objects;
 @RequiredArgsConstructor
 public class MileStoneServiceImpl implements MileStoneService {
 
+    private final ProjectMemberRepository projectMemberRepository;
     private final MileStoneRepository mileStoneRepository;
     private final ProjectRepository projectRepository;
 
@@ -41,14 +43,17 @@ public class MileStoneServiceImpl implements MileStoneService {
     }
 
     @Override
-    public MileStoneDto create(String title, LocalDate initDate, LocalDate dueDate, Long projectId) {
+    public MileStoneDto create(String userId, String title, LocalDate initDate, LocalDate dueDate, Long projectId) {
         if (Objects.isNull(title) || Objects.isNull(projectId) || title.isEmpty() || projectId <= 0) {
             throw new IllegalArgumentException("마일스톤에 필요한 값이 올바르지 않습니다.");
         }
-
+        if (!projectMemberRepository.existsByUserId(userId)) {
+            throw new ProjectMemberUserNotFoundException(userId);
+        }
         if (mileStoneRepository.existsByTitle(title)) {
             throw new MileStoneAlreadyExistsException();
         }
+
 
         Project project = projectRepository.findById(projectId)
                 .orElseThrow(ProjectNotFoundException::new);
@@ -59,26 +64,34 @@ public class MileStoneServiceImpl implements MileStoneService {
     }
 
     @Override
-    public void delete(Long mileStoneId) {
+    public void delete(Long mileStoneId, String userId) {
         if (Objects.isNull(mileStoneId) || mileStoneId <= 0) {
             throw new IllegalArgumentException("마일스톤 값이 올바르지 않습니다.");
+        }
+        if (!projectMemberRepository.existsByUserId(userId)) {
+            throw new ProjectMemberUserNotFoundException(userId);
         }
         if (!mileStoneRepository.existsById(mileStoneId)) {
             throw new MileStoneNotFoundException();
         }
+
 
         mileStoneRepository.delete(mileStoneRepository.findById(mileStoneId)
                                                         .orElseThrow(MileStoneNotFoundException::new));
     }
 
     @Override
-    public MileStoneDto update(Long mileStoneId, MileStoneRequest request) {
+    public MileStoneDto update(Long mileStoneId, MileStoneRequest request, String userId) {
         if (Objects.isNull(mileStoneId) || mileStoneId <= 0) {
             throw new IllegalArgumentException("마일스톤 값이 올바르지 않습니다.");
+        }
+        if (!projectMemberRepository.existsByUserId(userId)) {
+            throw new ProjectMemberUserNotFoundException(userId);
         }
         if (!mileStoneRepository.existsById(mileStoneId)) {
             throw new MileStoneNotFoundException();
         }
+
 
         MileStone mileStone = mileStoneRepository.findById(mileStoneId).orElseThrow(MileStoneNotFoundException::new);
         mileStone.setTitle(request.getTitle());
